backend/requirements.txt
# Flask and extensions
Flask==3.0.0
Flask-SocketIO==5.3.6
Flask-CORS==4.0.0
python-socketio==5.10.0
eventlet==0.34.2
gevent==23.9.1
gevent-websocket==0.10.1

# Database
SQLAlchemy==2.0.23

# HTTP client
requests==2.31.0
aiohttp==3.9.1
httpx==0.25.2

# Audio processing
pydub==0.25.1
soundfile==0.12.1
numpy==1.26.2

# Audio processing (audioop is built into Python 3.11)

# Environment and config
python-dotenv==1.0.0

# Async support
asyncio-mqtt==0.16.2

# Utilities
python-dateutil==2.8.2
pytz==2024.1

# Calendar parsing
icalendar==5.0.11

